### 2022-06-25 #4506 @editorialbot
[PRE REVIEW]: GlobalSensitivity.jl: julia package for performant global sensitivity analysis
**Submitting author:** @Vaibhavdixit02 (<a href="http://orcid.org/0000-0001-7763-2717">Vaibhav Dixit</a>)
**Repository:** https://github.com/SciML/GlobalSensitivity.jl
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
