### 2022-06-27 #4506 @Kevin-Mattheus-Moerman
@Vaibhavdixit02 @ChrisRackauckas thanks for this submission to JOSS. Below are some comments to work on before we proceed.

- [ ] Please check those potentially missing and invalid DOIs :point_up: you can update the paper by calling `@editorialbot generate pdf` and you can check the references again by calling `@editorialbot check references`.
- [ ] Please complete your affiliations to include city and country and do not use acronyms (e.g. spell our Massachusetts Institute of Technology, and Unites States of America), and if you'd like to add a department for an academic affiliation you can.
- [ ] Please extend your README to be more complete, i.e. have a short description/summary, provide information on documentation, installation, and getting started. In addition please add contributing guidelines. An example of a more complete one can be found here:
- https://github.com/gridap/Gridap.jl
- [ ] I found a poster discussing this work. The poster is fine, but can you comment on if this work has been published anywhere else?
- [ ] On the poster, it mentions an additional author `Vijay Ivaturi`, however this person is not listed on this JOSS submission. Can you confirm this is correct?
- [ ] Please can you add descriptive figure captions. Currently the captions appear to be short labels.
- [ ] The draft paper lacks an Acknowledgements section. This section is not a requirement but I am just pointing it out in case you'd like to add one.
