+++ joss/paper.bib
@@ -452,3 +452,14 @@ abstract = {Models are increasingly being relied upon to inform and support natu
   title = {Makie.jl: Flexible high-performance data visualization for Julia},
   journal = {Journal of Open Source Software}
 }
+
+@misc{ChristPlots2022,
+  doi = {10.48550/ARXIV.2204.08775},
+  url = {https://arxiv.org/abs/2204.08775},
+  author = {Christ, Simon and Schwabeneder, Daniel and Rackauckas, Christopher and Borregaard, Michael Krabbe and Breloff, Thomas},
+  keywords = {Graphics (cs.GR), FOS: Computer and information sciences, FOS: Computer and information sciences, I.3.3},
+  title = {Plots.jl -- a user extendable plotting API for the julia programming language},
+  publisher = {arXiv},
+  year = {2022},
+  copyright = {Creative Commons Attribution 4.0 International}
+}
