+++ joss/paper.bib
@@ -463,3 +463,14 @@ abstract = {Models are increasingly being relied upon to inform and support natu
   year = {2022},
   copyright = {Creative Commons Attribution 4.0 International}
 }
+
+@misc{RackauckasUDE,
+  doi = {10.48550/ARXIV.2001.04385},
+  url = {https://arxiv.org/abs/2001.04385},
+  author = {Rackauckas, Christopher and Ma, Yingbo and Martensen, Julius and Warner, Collin and Zubov, Kirill and Supekar, Rohit and Skinner, Dominic and Ramadhan, Ali and Edelman, Alan},
+  keywords = {Machine Learning (cs.LG), Dynamical Systems (math.DS), Quantitative Methods (q-bio.QM), Machine Learning (stat.ML), FOS: Computer and information sciences, FOS: Computer and information sciences, FOS: Mathematics, FOS: Mathematics, FOS: Biological sciences, FOS: Biological sciences},
+  title = {Universal Differential Equations for Scientific Machine Learning},
+  publisher = {arXiv},
+  year = {2020},
+  copyright = {arXiv.org perpetual, non-exclusive license}
+}
