### 2022-06-27 7a12640
A bunch of paper edits for the review
https://github.com/openjournals/joss-reviews/issues/4506#issuecomment-1167028890
