+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'GlobalSensitivity.jl: julia package for performant global sensitivity analysis'
+title: 'GlobalSensitivity.jl: Performant and Parallel Global Sensitivity Analysis with Julia'
 tags:
   - julia
   - global sensitivity analysis
@@ -9,12 +9,14 @@ authors:
     corresponding: true
     affiliation: 1
   - name: Christopher Rackauckas
-    affiliation: "1, 2"
+    affiliation: "1, 2, 3"
 affiliations:
  - name: Julia Computing
    index: 1
- - name: MIT
+ - name: Massachusetts Institute of Technology
    index: 2
+ - name: Pumas-AI
+   index: 3
 date: 21 June 2022
 bibliography: paper.bib
 
@@ -25,23 +27,23 @@ bibliography: paper.bib
 Global Sensitivity Analysis (GSA) methods are used to quantify the uncertainty in
 output of a model with respect to the parameters. These methods allow practitioners to
 measure both parameter's individual contributions and the contribution of their interactions
-to the output uncertainity. GlobalSensitivity.jl is a julia [@Bezanson2017] package containing implementation of some the most popular GSA methods. Currently it supports Delta Moment-Independent [@Borgonovo2007,@Plischke2013], DGSM [@Sobol2009], EASI [@Plischke2010] [@Plischke2012], eFAST [@Saltelli1999] [@Saltelli1998], Morris [@Morris1991] [@Campolongo2007], Fractional Factorial [@Saltelli2008b] , RBD-FAST [@Tarantola2006], Sobol [@Saltelli2008b] [@Sobol2001] [@Saltelli2002a] and Regression based sensitivity [@Guido2016] methods.
+to the output uncertainity. GlobalSensitivity.jl is a Julia [@Bezanson2017] package containing implementation of some the most popular GSA methods. Currently it supports Delta Moment-Independent [@Borgonovo2007,@Plischke2013], DGSM [@Sobol2009], EASI [@Plischke2010] [@Plischke2012], eFAST [@Saltelli1999] [@Saltelli1998], Morris [@Morris1991] [@Campolongo2007], Fractional Factorial [@Saltelli2008b] , RBD-FAST [@Tarantola2006], Sobol [@Saltelli2008b] [@Sobol2001] [@Saltelli2002a] and Regression based sensitivity [@Guido2016] methods.
 
 # Statement of need
 
-Global Sensitivity Analysis has become an essential part of modeling workflow for practitioners in various fields such as Quantitative Systems Pharmacology and Environmental Modeling [@saltelli2020five] [@JAKEMAN2006602] [@sher2022quantitative] [@zhang2015sobol]. It can be used primarily in two stages, either before parameter estimation to simplify the fitting problem by fixing unimportant parameters or for analysis of the input parameters' influence on the output.
+Global Sensitivity Analysis has become an essential part of modeling workflows for practitioners in various fields such as Quantitative Systems Pharmacology and Environmental Modeling [@saltelli2020five] [@JAKEMAN2006602] [@sher2022quantitative] [@zhang2015sobol]. It can be used primarily in two stages, either before parameter estimation to simplify the fitting problem by fixing unimportant parameters or for analysis of the input parameters' influence on the output.
 
-GlobalSensitivity.jl provides julia implementation of some of the popular GSA methods. Thus it benefits from the performance advantage of julia, provides a convenient unified API for different GSA methods by leveraging multiple dispatch and has a parallelized implementation for some of the methods.
+GlobalSensitivity.jl provides a Julia implementation of some of the popular GSA methods. Thus it benefits from the performance advantage of Julia, provides a convenient unified API for different GSA methods by leveraging multiple dispatch and has a parallelized implementation for some of the methods.
 
[…]
