### 2022-06-27 #4506 @ChrisRackauckas
>  I found a poster discussing this work. The poster is fine, but can you comment on if this work has been published anywhere else?

No prior publication

>  On the poster, it mentions an additional author Vijay Ivaturi, however this person is not listed on this JOSS submission. Can you confirm this is correct?

This is correct.

> gsa function should read GSA function

This should not because that would be a syntax error as functions in Julia need to respect capitalization. It's `gsa` function, so the `gsa` is highlighted as syntax rather than prose, and thus it should be matching capitalization.
