+++ joss/paper.bib
@@ -323,7 +323,7 @@ number = {4},
 pages = {354-360},
 year = {2010},
 issn = {0951-8320},
-doi = {https://doi.org/10.1016/j.ress.2009.11.005},
+doi = {10.1016/j.ress.2009.11.005},
 url = {https://www.sciencedirect.com/science/article/pii/S0951832009002579},
 author = {Elmar Plischke},
 keywords = {Global sensitivity analysis, Sobol’ sensitivity indices, FAST method, Correlation ratio, Post-processing algorithm, Space filling curves},
@@ -337,7 +337,7 @@ volume = {35},
 pages = {188-191},
 year = {2012},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2012.03.004},
+doi = {10.1016/j.envsoft.2012.03.004},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815212000862},
 author = {Elmar Plischke},
 keywords = {Variance-based sensitivity indicators, First order effects, Discrete cosine transformation},
@@ -402,7 +402,7 @@ number = {5},
 pages = {602-614},
 year = {2006},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2006.01.004},
+doi = {10.1016/j.envsoft.2006.01.004},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815206000107},
 author = {A.J. Jakeman and R.A. Letcher and J.P. Norton},
 keywords = {Model testing, Verification, Uncertainty, Sensitivity, Integrated assessment, System identification},
@@ -416,6 +416,7 @@ abstract = {Models are increasingly being relied upon to inform and support natu
   volume={84},
   number={3},
   pages={1--15},
+  doi={10.1007/s11538-021-00982-5},
   year={2022},
   publisher={Springer}
 }
@@ -428,6 +429,7 @@ abstract = {Models are increasingly being relied upon to inform and support natu
   number={2},
   pages={69--79},
   year={2015},
+  doi={10.1002/psp4.6},
   publisher={Wiley Online Library}
 }
 
