### 2022-07-04 06c179f
Add examples and plots in readme
