### 2022-07-04 465b0b8
CompatHelper: add new compat entry for CairoMakie at version 0.8, (keep existing compat)
