### 2022-07-04 ce73e8f
Add headings to plots in paper
