+++ joss/paper.md
@@ -68,7 +68,7 @@ end
 
 bounds = [[1,5],[1,5],[1,5],[1,5]]
 
-reg_sens = gsa(f1, RegressionGSA(true), bounds)
+reg_sens = gsa(f1, RegressionGSA(true), bounds; samples = 1000)
 fig = Figure(resolution = (600, 400))
 ax, hm = CairoMakie.heatmap(fig[1,1], reg_sens.partial_correlation,
                             figure = (resolution = (300, 200),),
@@ -89,6 +89,8 @@ Colorbar(fig[2, 2], hm)
 fig
 ```
 
+#### Regression based GSA for Lotka Volterra model
+
 ![heatmapreg](https://user-images.githubusercontent.com/23134958/127019339-607b8d0b-6c38-4a18-b62e-e3ea0ae40941.png)
 
 Next, the Morris method is used and results are visualized as a scatter plot.
@@ -107,13 +109,15 @@ scatter(fig[1,2], [1,2,3,4], morris_sens.means_star[2,:],
 fig
 ```
 
+#### Morris method GSA for Lotka Volterra model
+
 ![morrisscat](https://user-images.githubusercontent.com/23134958/127019346-2b5548c5-f4ec-4547-9f8f-af3e4b4c317c.png)
 
 Here we show use of the Sobol and eFAST methods, the first order and total order indices are plotted for both the dependent variables for all four parameters.
 
 ```julia
-sobol_sens = gsa(f1, Sobol(), bounds, N=5000)
-efast_sens = gsa(f1, eFAST(), bounds)
+sobol_sens = gsa(f1, Sobol(), bounds, samples=5000)
+efast_sens = gsa(f1, eFAST(), bounds, samples=5000)
 fig = Figure(resolution = (300, 200))
 barplot(fig[1,1], [1,2,3,4], sobol_sens.S1[1, :],
         color = :green, axis = (xticksvisible = false,
@@ -146,6 +150,8 @@ barplot(fig[2,2], [1,2,3,4], efast_sens.ST[2, :],
         xticklabelsvisible = false))
 fig
 ```
+
+#### Sobol and eFAST methods GSA for Lotka Volterra model
 
 ![sobolefastprey](https://user-images.githubusercontent.com/23134958/127019361-8d625107-7f9c-44b5-a0dc-489bd512b7dc.png)
 ![sobolefastpred](https://user-images.githubusercontent.com/23134958/127019358-8bd0d918-e6fd-4929-96f1-d86330d91c69.png)
@@ -192,7 +198,7 @@ f1 = function (p)
            prob1 = remake(prob;p=p)
            sol = solve(prob1,Tsit5();saveat=t)
        end
-sobol_sens = gsa(f1, Sobol(nboot = 20), bounds, N=5000)
+sobol_sens = gsa(f1, Sobol(nboot = 20), bounds, samples=5000)
 fig = Figure(resolution = (300, 200))
 ax, hm = CairoMakie.scatter(fig[1,1], sobol_sens.S1[1][1,2:end],
                             label = "Prey", markersize = 4)
@@ -219,6 +225,8 @@ CairoMakie.scatter!(fig[2,2], sobol_sens.S1[4][2,2:end],
 title = Label(fig[0,:], "First order Sobol indices")
 legend = Legend(fig[2,3], ax)
 ```
+
+#### Sobol method GSA on timeseries of Lotka Volterra model
 
 ![timeseriessobollv](https://user-images.githubusercontent.com/23134958/156987652-85958bde-ae73-4f71-8555-318f779257ad.png)
 
