### 2022-07-04 #4506 @ChrisRackauckas
It's still missing figure captions. @Vaibhavdixit02 can you take those on? I think that's the last piece.
