### 2022-07-05 df493c0
Make captions bold text below the plots
