### 2022-07-05 d5e71ba
Make captions in the plot itself
