+++ joss/paper.md
@@ -86,11 +86,11 @@ ax, hm = CairoMakie.heatmap(fig[2,1], reg_sens.standard_regression,
                             xticklabelsvisible = false,
                             title = "Standard regression"))
 Colorbar(fig[2, 2], hm)
+Label(fig[3, :], "Regression based GSA for Lotka Volterra model")
 fig
 ```
 
-![heatmapreg](https://user-images.githubusercontent.com/23134958/127019339-607b8d0b-6c38-4a18-b62e-e3ea0ae40941.png)
-**Regression based GSA for Lotka Volterra model**
+![heatmapreg](https://user-images.githubusercontent.com/23134958/177251367-9c93a219-8871-4734-8e3f-b6e3f6b6d90a.png)
 
 Next, the Morris method is used and results are visualized as a scatter plot.
 
@@ -98,25 +98,25 @@ Next, the Morris method is used and results are visualized as a scatter plot.
 using StableRNGs
 _rng = StableRNG(1234)
 morris_sens = gsa(f1, Morris(), bounds, rng = _rng)
-fig = Figure(resolution = (300, 200))
+fig = Figure(resolution = (400, 300))
 scatter(fig[1,1], [1,2,3,4], morris_sens.means_star[1,:],
         color = :green, axis = (xticksvisible = false,
         xticklabelsvisible = false, title = "Prey (Morris)",))
 scatter(fig[1,2], [1,2,3,4], morris_sens.means_star[2,:],
         color = :red, axis = (xticksvisible = false,
         xticklabelsvisible = false, title = "Predator (Morris)",))
+Label(fig[2, :], "Morris method GSA for Lotka Volterra model")
 fig
 ```
 
-![morrisscat](https://user-images.githubusercontent.com/23134958/127019346-2b5548c5-f4ec-4547-9f8f-af3e4b4c317c.png)
-**Morris method GSA for Lotka Volterra model**
+![morrisscat](https://user-images.githubusercontent.com/23134958/177251158-69a6e8b8-0a46-4998-b9f2-106886827176.png)
 
 Here we show use of the Sobol and eFAST methods, the first order and total order indices are plotted for both the dependent variables for all four parameters.
 
 ```julia
 sobol_sens = gsa(f1, Sobol(), bounds, samples=5000)
 efast_sens = gsa(f1, eFAST(), bounds, samples=5000)
-fig = Figure(resolution = (300, 200))
+fig = Figure(resolution = (600, 400))
 barplot(fig[1,1], [1,2,3,4], sobol_sens.S1[1, :],
         color = :green, axis = (xticksvisible = false,
         xticklabelsvisible = false, title = "Prey (Sobol)",
@@ -130,9 +130,10 @@ barplot(fig[1,2], [1,2,3,4], efast_sens.S1[1, :],
 barplot(fig[2,2], [1,2,3,4], efast_sens.ST[1, :],
         color = :red, axis = (xticksvisible = false,
         xticklabelsvisible = false))
+Label(fig[3,:], "Sobol and eFAST methods GSA for mean of Prey population")
 fig
 
-fig = Figure(resolution = (300, 200))
+fig = Figure(resolution = (600, 400))
 barplot(fig[1,1], [1,2,3,4], sobol_sens.S1[2, :],
         color = :green, axis = (xticksvisible = false,
         xticklabelsvisible = false, title = "Predator (Sobol)",
@@ -146,12 +147,12 @@ barplot(fig[1,2], [1,2,3,4], efast_sens.S1[2, :],
 barplot(fig[2,2], [1,2,3,4], efast_sens.ST[2, :],
         color = :red, axis = (xticksvisible = false,
[…]
