### 2022-07-07 d8c29ec
Update joss/paper.md
Co-authored-by: Christopher Rackauckas <accounts@chrisrackauckas.com>
