+++ joss/paper.md
@@ -90,7 +90,7 @@ Label(fig[3, :], "Regression based GSA for Lotka Volterra model")
 fig
 ```
 
-![Heatmap of partial correlation and standard regression coefficients for the maximum of the prey population and the mean of the predator population from the Lotka Volterra model. The first column in each of the heatmaps corresponds to the maximum of the predator population and the second column corresponds the mean of the prey population. The rows follow the order of the parameters in the vector, first row being `p[1]` second `p[2]` and so on.](https://user-images.githubusercontent.com/23134958/177251367-9c93a219-8871-4734-8e3f-b6e3f6b6d90a.png)
+![**Heatmap of partial correlation and standard regression coefficients for the maximum of the prey population and the mean of the predator population.** The first column in both the heatmaps corresponds to the maximum of the predator population and the second column corresponds the mean of the prey population from the Lotka-Volterra ODE solved on the time interval `(0.0,10.0)`. The rows follow the order of the parameters in the vector, first row being `p[1]` second `p[2]` and so on.](https://user-images.githubusercontent.com/23134958/177251367-9c93a219-8871-4734-8e3f-b6e3f6b6d90a.png)
 
 Next, the Morris method is used and results are visualized as a scatter plot.
 
@@ -101,15 +101,15 @@ morris_sens = gsa(f1, Morris(), bounds, rng = _rng)
 fig = Figure(resolution = (400, 300))
 scatter(fig[1,1], [1,2,3,4], morris_sens.means_star[1,:],
         color = :green, axis = (xticksvisible = false,
-        xticklabelsvisible = false, title = "Prey (Morris)",))
+        xticklabelsvisible = false, title = "(a) Prey",))
 scatter(fig[1,2], [1,2,3,4], morris_sens.means_star[2,:],
         color = :red, axis = (xticksvisible = false,
-        xticklabelsvisible = false, title = "Predator (Morris)",))
+        xticklabelsvisible = false, title = "(b) Predator",))
 Label(fig[2, :], "Morris method GSA for Lotka Volterra model")
 fig
 ```
 
-![**Scatter plot of `μ*` for the mean of the prey population and maximum of the predator population from the Morris method.** Depicted are the Morris indices for the mean of the time series solutions for the Lotka-Volterra ODE solved on the time interval `(0.0,10.0)` with parameters as demonstrated in the code sample. Each of the four parameters occur in both plots as one scatter point with `μ*` on the y-axis and index of the parameter corresponding to its order on the x-axis. Left: the Morris indices for the mean of the prey population. Right: the Morris indices for the maximum of the predator population.](https://user-images.githubusercontent.com/23134958/177251158-69a6e8b8-0a46-4998-b9f2-106886827176.png)
+![**Scatter plot of `μ*` for the mean of the prey population and maximum of the predator population from the Morris method.** Depicted are the Morris method's mean of the distribution of the absolute values of the elementary effects of the input factors for the mean of the time series solution for the first dependent variable (prey population) and maximum of the second dependent variable (predator population) for the Lotka-Volterra ODE solved on the time interval `(0.0,10.0)` with parameters as demonstrated in the code sample. Each of the four parameters occur in both plots as one scatter point with `μ*` on the y-axis and index of the parameter corresponding to its order on the x-axis. (a): the Morris indices for the mean of the prey population. (b): the Morris indices for the maximum of the predator population.](https://user-images.githubusercontent.com/23134958/177814529-9eda8dfb-b2ae-41da-94a9-d704a24f84d2.png)
 
 Here we show use of the Sobol and eFAST methods, the first order and total order indices are plotted for both the dependent variables for all four parameters.
 
@@ -151,8 +151,8 @@ Label(fig[3,:], "Sobol and eFAST methods GSA for maximum of Predator population"
 fig
 ```
 
[…]
