### 2022-07-07 #4506 @alstat
Hi @jbytecode, I'll pass, I don't have the expertise on differential equations. Thanks.

> 👋👋👋 Dear @alstat 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**GlobalSensitivity.jl: julia package for performant global sensitivity analysis**". You can find more information at the top of this Github issue (#4506).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
