### 2022-07-08 #4506 @jbytecode
👋👋👋 Dear @mkitti 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**GlobalSensitivity.jl: julia package for performant global sensitivity analysis**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4506).
