### 2022-07-09 #4506 @storyetfall
> @storyetfall - I have just seen that you responded with an emoji to our invitation. Is that mean you accept the invitation?

Sorry for the vagueness. Indeed -- I am happy to take a look.
