### 2022-07-09 #4561 @editorialbot
[REVIEW]: GlobalSensitivity.jl: julia package for performant global sensitivity analysis
**Submitting author:** @Vaibhavdixit02 (<a href="http://orcid.org/0000-0001-7763-2717">Vaibhav Dixit</a>)
**Repository:** https://github.com/SciML/GlobalSensitivity.jl
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.1
