### 2022-07-29 SciML/GlobalSensitivity.jl#68 @zhenwu0728
[Documentation] Docstring for gsa function missing
It would be nice to add docstring for gsa function and include the argument lists for different methods.

Currently, the docs are like:

### API

`function gsa(f, method::Morris, p_range::AbstractVector; batch=false, kwargs...)`

Something like [this](https://clima.github.io/OceananigansDocumentation/stable/appendix/library/#Models) would be better.

[openjournals/joss-reviews#4561](https://github.com/openjournals/joss-reviews/issues/4561)
