### 2022-07-29 SciML/GlobalSensitivity.jl#69 @zhenwu0728
[Documentation] Tutorial doesn't work
The tutorial for [Global Sensitivity Analysis of the Lotka-Volterra model](https://gsa.sciml.ai/stable/tutorials/juliacon21/#Global-Sensitivity-Analysis-of-the-Lotka-Volterra-model) run into errors. The argument `samples` is missing in gsa function.

Please double-check the two tutorials.

[openjournals/joss-reviews#4561](https://github.com/openjournals/joss-reviews/issues/4561)
