### 2022-07-29 #4561 @zhenwu0728
Hi @jbytecode, I would like to submit my review.

Overall, the package looks good to me, but I still have some comments:
- I suggest removing the Example section in the paper.
- Multiple citations in one place should be in the same bracket.
- I suggest removing some of the citations in Summary - one per method should be enough.
- Lack of docstring for gas function ([Issue#68](https://github.com/SciML/GlobalSensitivity.jl/issues/68))
- Tutorial doesn't work ([Issue#69](https://github.com/SciML/GlobalSensitivity.jl/issues/69))

All of these remarks should be easy to fix.
