### 2022-08-01 1b3f0c6
Update tutorials and switch to using docstrings
