+++ src/sobol_sensitivity.jl
@@ -3,6 +3,16 @@ struct Sobol <: GSAMethod
     nboot::Int
     conf_level::Float64
 end
+
+"""
+    Sobol(; order = [0, 1], nboot = 1, conf_level = 0.95)
+
+- order: The order of the indices to calculate. Defaults to [0,1], which means the Total
+         and First order indices. Passing 2 enables calculation of the Second order indices as well.
+
+For confidence interval calculation nboot should be specified for the number (>0) of bootstrap
+runs and conf_level for the confidence level, the default for which is 0.95.
+"""
 Sobol(; order = [0, 1], nboot = 1, conf_level = 0.95) = Sobol(order, nboot, conf_level)
 
 mutable struct SobolResult{T1, T2, T3, T4}
@@ -241,6 +251,22 @@ function gsa_sobol_all_y_analysis(method, all_y::AbstractArray{T}, d, n, Ei_esti
                        nboot > 1 ? reshape(ST_CI, size_...) : nothing)
 end
 
+"""
+    gsa(f, method::Sobol, p_range::AbstractVector; samples, kwargs...)
+    gsa(f, method::Sobol, A::AbstractMatrix{TA}, B::AbstractMatrix;
+             batch = false, Ei_estimator = :Jansen1999,
+             distributed::Val{SHARED_ARRAY} = Val(false),
+             kwargs...) where {TA, SHARED_ARRAY}
+
+`Ei_estimator` can take `:Homma1996`, `:Sobol2007` and `:Jansen1999` for which
+  Monte Carlo estimator is used for the `Ei` term. Defaults to `:Jansen1999`. Details for these can be found in the
+  corresponding papers:
+
+    - `:Homma1996` - [Homma, T. and Saltelli, A., 1996. Importance measures in global sensitivity analysis of nonlinear models. Reliability Engineering & System Safety, 52(1), pp.1-17.](https://www.sciencedirect.com/science/article/abs/pii/0951832096000026)
+    - `:Sobol2007` - [I.M. Sobol, S. Tarantola, D. Gatelli, S.S. Kucherenko and W. Mauntz, 2007, Estimating the approx- imation errors when fixing unessential factors in global sensitivity analysis, Reliability Engineering and System Safety, 92, 957–960.](https://www.sciencedirect.com/science/article/abs/pii/S0951832006001499)
+    [A. Saltelli, P. Annoni, I. Azzini, F. Campolongo, M. Ratto and S. Tarantola, 2010, Variance based sensitivity analysis of model output. Design and estimator for the total sensitivity index, Computer Physics Communications 181, 259–270.](https://www.sciencedirect.com/science/article/abs/pii/S0010465509003087)
+    - `:Jansen1999` - [M.J.W. Jansen, 1999, Analysis of variance designs for model output, Computer Physics Communi- cation, 117, 35–43.](https://www.sciencedirect.com/science/article/abs/pii/S0010465598001544)
+"""
 function gsa(f, method::Sobol, p_range::AbstractVector; samples, kwargs...)
     AB = QuasiMonteCarlo.generate_design_matrices(samples, [i[1] for i in p_range],
                                                   [i[2] for i in p_range],
