### 2022-08-01 677016d
Fix code blocks formatting and move details and examples into docstrings
