+++ previews/PR70/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"methods/regression/#Regression-Method","page":"Regression Method","title":"Regression Method","text":"","category":"section"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"RegressionGSA(; rank::Bool = false)","category":"page"},{"location":"methods/regression/#GlobalSensitivity.RegressionGSA-Tuple{}","page":"Regression Method","title":"GlobalSensitivity.RegressionGSA","text":"RegressionGSA(; rank::Bool = false)\n\nRegressionGSA methods for global sensitivity analysis. Providing this to gsa results in a calculation of the following statistics, provided as a RegressionGSAResult. If the function f to be analyzed is of dimensionality f: R^n -> R^m, then these coefficients are returned as a matrix, with the corresponding statistic in the (i, j) entry.\n\npearson: This is equivalent to the correlation coefficient matrix between input and\n\noutput. The rank version is known as the Spearman coefficient.\n\nstandard_regression: Standard regression coefficients, also known as sigma-normalized\n\nderivatives\n\npartial_correlation: Partial correlation coefficients, related to the precision matrix\n\nand a measure of the correlation of linear models of the\n\nArguments\n\nrank::Bool = false: Flag determining whether to also run a rank regression analysis\n\n\n\n\n\n","category":"method"},{"location":"methods/regression/#Regression-Details","page":"Regression Method","title":"Regression Details","text":"","category":"section"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"It is possible to fit a linear model explaining the behavior of Y given the values of X, provided that the sample size n is sufficiently large (at least n > d).","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"The measures provided for this analysis by us in GlobalSensitivity.jl are","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"a) Pearson Correlation Coefficient:","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"r = fracsum_i=1^n (x_i - overlinex)(y_i - overliney)sqrtsum_i=1^n (x_i - overlinex)^2(y_i - overliney)^2","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"b) Standard Regression Coefficient (SRC):","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"SRC_j = beta_j sqrtfracVar(X_j)Var(Y)","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"where beta_j is the linear regression coefficient associated to X_j. This is also known as a sigma-normalized derivative.","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"c) Partial Correlation Coefficient (PCC):","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"PCC_j = rho(X_j - hatX_-jY_j - hatY_-j)","category":"page"},{"location":"methods/regression/","page":"Regression Method","title":"Regression
[…]
