### 2022-08-04 b77ff36
Move everything to `struct`/constructor docstrings
