+++ src/regression_sensitivity.jl
@@ -1,8 +1,5 @@
-struct RegressionGSA <: GSAMethod
-    rank::Bool
-end
-
 @doc raw"""
+
     RegressionGSA(; rank::Bool = false)
 
 RegressionGSA methods for global sensitivity analysis. Providing this to `gsa` results
@@ -20,7 +17,7 @@ and a measure of the correlation of linear models of the
 # Arguments
 - `rank::Bool = false`: Flag determining whether to also run a rank regression analysis
 
-## Regression Method Details
+## Method Details
 
 It is possible to fit a linear model explaining the behavior of Y given the
 values of X, provided that the sample size n is sufficiently large (at least n > d).
@@ -54,7 +51,44 @@ linear model where ``X_j`` is absent. PCC measures the sensitivity of ``Y`` to
 ``X_j`` when the effects of the other inputs have been canceled.
 
 If `rank` is set to `true`, then the rank coefficients are also calculated.
+
+## API
+
+    gsa(f, method::RegressionGSA, p_range::AbstractVector; samples::Int, batch = false)
+    gsa(X, Y, method::RegressionGSA)
+
+
+### Example
+
+```julia
+using GlobalSensitivity
+
+function linear_batch(X)
+    A= 7
+    B= 0.1
+    @. A*X[1,:]+B*X[2,:]
+end
+function linear(X)
+    A= 7
+    B= 0.1
+    A*X[1]+B*X[2]
+end
+
+p_range = [[-1, 1], [-1, 1]]
+reg = gsa(linear_batch, RegressionGSA(), p_range; batch = true)
+
+reg = gsa(linear, RegressionGSA(), p_range; batch = false)
+reg = gsa(linear, RegressionGSA(true), p_range; batch = false) #with rank coefficients
+
+X = QuasiMonteCarlo.sample(1000, [-1, -1], [1, 1], QuasiMonteCarlo.SobolSample())
+Y = reshape(linear.([X[:, i] for i in 1:1000]), 1, 1000)
+reg_mat = gsa(X, Y, RegressionGSA(true))
+```
 """
+struct RegressionGSA <: GSAMethod
+    rank::Bool
+end
+
 RegressionGSA(; rank::Bool = false) = RegressionGSA(rank)
 
 struct RegressionGSAResult{T, TR}
@@ -113,38 +147,6 @@ function _calculate_partial_correlation_coefficients(X, Y)
     return Matrix(transpose(pcc_XY[axes(X, 1), lastindex(X, 1) .+ axes(Y, 1)]))
 end
 
-"""
-    gsa(f, method::RegressionGSA, p_range::AbstractVector; samples::Int, batch = false)
-    gsa(X, Y, method::RegressionGSA)
-
-
-### Example
-
-```julia
-using GlobalSensitivity
-
-function linear_batch(X)
-    A= 7
-    B= 0.1
-    @. A*X[1,:]+B*X[2,:]
-end
-function linear(X)
-    A= 7
-    B= 0.1
-    A*X[1]+B*X[2]
-end
-
-p_range = [[-1, 1], [-1, 1]]
-reg = gsa(linear_batch, RegressionGSA(), p_range; batch = true)
-
-reg = gsa(linear, RegressionGSA(), p_range; batch = false)
-reg = gsa(linear, RegressionGSA(true), p_range; batch = false) #with rank coefficients
-
-X = QuasiMonteCarlo.sample(1000, [-1, -1], [1, 1], QuasiMonteCarlo.SobolSample())
-Y = reshape(linear.([X[:, i] for i in 1:1000]), 1, 1000)
-reg_mat = gsa(X, Y, RegressionGSA(true))
-```
-"""
[…]
