+++ docs/src/methods/sobol.md
@@ -1,11 +1,5 @@
 # Sobol Method
 
 ```@docs
-Sobol(; order = [0, 1], nboot = 1, conf_level = 0.95)
-```
-
-## API
-
-```@docs
-gsa(f, method::Sobol, p_range::AbstractVector; samples, kwargs...)
+Sobol
 ```
