+++ docs/src/methods/rbdfast.md
@@ -1,11 +1,5 @@
 # Random Balance Design FAST Method
 
 ```@docs
-RBDFAST(; num_harmonics = 6)
-```
-
-## API
-
-```@docs
-gsa(f, method::RBDFAST; num_params, samples, rng::AbstractRNG = Random.default_rng(), batch = false, kwargs...)
+RBDFAST
 ```

[…]
