### 2022-08-04 #4561 @zhenwu0728
>

> Thanks for the review and also opening the issues @zhenwu0728, that was very helpful.
>
> > I suggest removing the Example section in the paper.
>
> I think it makes sense, I will describe the experiment/example in text and add a link to the docs' section that has this example.
>
> > I suggest removing some of the citations in Summary - one per method should be enough.
>
> They are there because of subsequent modifications offered by those papers and also the package contains multiple implementations for those papers.
>
> Let me know what you think?

Both sound good to me.
