### 2022-08-04 91dfe1c
Remove example and combine multiple citations into one bracket in joss
