### 2022-08-05 #4561 @jbytecode
citations look good now.

one more issue: giving mutable links in pdf is a little bit problematic if you decide to change folder structure and the content in future. The links `1` and `2` route to changeable addresses by time.

The solution: can you provide a short example that presents the basic functionality of the software rather than a three-pages long one in manuscript?

@zhenwu0728 - how do you think about that?
