### 2022-08-08 3ca4f01
Make docs build examples strictly
