### 2022-08-08 7779d3f
Update parallelized_gsa.md
