+++ docs/Project.toml
@@ -8,4 +8,4 @@ QuasiMonteCarlo = "8a4e6c94-4038-4cdc-81c3-7e6ffdb2a71b"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
-Documenter = "0.25.3"
+Documenter = "0.27"
