### 2022-08-09 #4561 @storyetfall
Review done! Everything looks good. Only a few very minor comments:

* several formatting issues in the method docs, including but not limited to:
    *  in Regression Method "R^n -> R^m" appears as plain text and not LaTeX.
    * in Regression Method there are newline issues in the first bullet points
    * in eFAST "samples" should be LaTeX'd as `\text{samples}` or replaced with a variable like `N_{s}` ?
    * in eFAST "num_harmonics" isn't described using the bullet point format of the other methods
    * in Fractional Factorial there is no footprint line at the top
* could be useful to explicitly define the Lotka-Volterra model in the text of the tutorials. it's clear by looking at the code and also many will be familiar, but there's no harm in being extra clear :)
