### 2022-08-11 1366c9b
Add StableRNGs and missing samples kwarg
