+++ src/sobol_sensitivity.jl
@@ -2,11 +2,9 @@
 
     Sobol(; order = [0, 1], nboot = 1, conf_level = 0.95)
 
-- `order`: The order of the indices to calculate. Defaults to [0,1], which means the Total
-         and First order indices. Passing 2 enables calculation of the Second order indices as well.
-
-For confidence interval calculation `nboot` should be specified for the number (>0) of bootstrap
-runs and `conf_level` for the confidence level, the default for which is 0.95.
+- `order`: the order of the indices to calculate. Defaults to [0,1], which means the Total and First order indices. Passing 2 enables calculation of the Second order indices as well.
+- `nboot`: for confidence interval calculation `nboot` should be specified for the number (>0) of bootstrap runs.
+- `conf_level`: the confidence level, the default for which is 0.95.
 
 ## Method Details
 
