### 2022-08-11 d049481
Fix plotting kwargs
