### 2022-08-13 fea0bae
Update docs/src/tutorials/juliacon21.md
