+++ docs/src/tutorials/juliacon21.md
@@ -27,7 +27,7 @@ end
 
 bounds = [[1,5],[1,5],[1,5],[1,5]]
 
-reg_sens = gsa(f1, RegressionGSA(true), bounds, samples = 1000)
+reg_sens = gsa(f1, RegressionGSA(true), bounds, samples = 200)
 fig = Figure(resolution = (600, 400))
 ax, hm = CairoMakie.heatmap(fig[1,1], reg_sens.partial_correlation, axis = (xticksvisible = false,yticksvisible = false, yticklabelsvisible = false, xticklabelsvisible = false, title = "Partial correlation"))
 Colorbar(fig[1, 2], hm)
