### 2022-08-13 1bf3641
Update parallelized_gsa.md
