### 2022-08-13 60a675b
Update parallelized_gsa.md
