+++ previews/PR73/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"methods/regression/#Regression-Method","page":"Regression Method","title":"Regression Method","text":"","category":"section"},{"location":"methods/regression/","page":"Regression Method","title":"Regression Method","text":"RegressionGSA","category":"page"},{"location":"methods/regression/#GlobalSensitivity.RegressionGSA","page":"Regression Method","title":"GlobalSensitivity.RegressionGSA","text":"RegressionGSA(; rank::Bool = false)\n\nrank::Bool = false: Flag determining whether to also run a rank regression analysis\n\nProviding this to gsa results in a calculation of the following statistics, provided as a RegressionGSAResult. If the function f to be analyzed is of dimensionality f R^n - R^m, then these coefficients are returned as a matrix, with the corresponding statistic in the (i, j)` entry.\n\npearson: This is equivalent to the correlation coefficient matrix between input and output. The rank version is known as the Spearman coefficient.\nstandard_regression: Standard regression coefficients, also known as sigma-normalized derivatives\npartial_correlation: Partial correlation coefficients, related to the precision matrix and a measure of the correlation of linear models of the\n\nMethod Details\n\nIt is possible to fit a linear model explaining the behavior of Y given the values of X, provided that the sample size n is sufficiently large (at least n > d).\n\nThe measures provided for this analysis by us in GlobalSensitivity.jl are\n\na) Pearson Correlation Coefficient:\n\nr = fracsum_i=1^n (x_i - overlinex)(y_i - overliney)sqrtsum_i=1^n (x_i - overlinex)^2(y_i - overliney)^2\n\nb) Standard Regression Coefficient (SRC):\n\nSRC_j = beta_j sqrtfracVar(X_j)Var(Y)\n\nwhere beta_j is the linear regression coefficient associated to X_j. This is also known as a sigma-normalized derivative.\n\nc) Partial Correlation Coefficient (PCC):\n\nPCC_j = rho(X_j - hatX_-jY_j - hatY_-j)\n\nwhere hatX_-j is the prediction of the linear model, expressing X_j with respect to the other inputs and hatY_-j is the prediction of the linear model where X_j is absent. PCC measures the sensitivity of Y to X_j when the effects of the other inputs have been canceled.\n\nIf rank is set to true, then the rank coefficients are also calculated.\n\nAPI\n\ngsa(f, method::RegressionGSA, p_range::AbstractVector; samples::Int, batch = false)\ngsa(X, Y, method::RegressionGSA)\n\nExample\n\nusing GlobalSensitivity\n\nfunction linear_batch(X)\n    A= 7\n    B= 0.1\n    @. A*X[1,:]+B*X[2,:]\nend\nfunction linear(X)\n    A= 7\n    B= 0.1\n    A*X[1]+B*X[2]\nend\n\np_range = [[-1, 1], [-1, 1]]\nreg = gsa(linear_batch, RegressionGSA(), p_range; batch = true)\n\nreg = gsa(linear, RegressionGSA(), p_range; batch = false)\nreg = gsa(linear, RegressionGSA(true), p_range; batch = false) #with rank coefficients\n\nX = QuasiMonteCarlo.sample(1000, [-1, -1], [1, 1], QuasiMonteCarlo.SobolSample())\nY = reshape(linear.([X[:, i] for i in 1:1000]), 1, 1000)\nreg_mat = gsa(X, Y, RegressionGSA(true))\n\n\n\n\n\n","category":"type"},{"location":"tutorials/parallelized_gsa/#Parallelized-Morris-and-Sobol-Sensitivity-Analysis-of-an-ODE","page":"
[…]
