+++ versions.js
@@ -7,3 +7,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
+var DOCUMENTER_NEWEST = "v2.0.1";
+var DOCUMENTER_STABLE = "stable";
