+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v2.1",
   "v2.0",
   "v1.3",
   "v1.2",
@@ -7,5 +8,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v2.0.1";
+var DOCUMENTER_NEWEST = "v2.1.0";
 var DOCUMENTER_STABLE = "stable";
