### 2022-08-15 #4561 @jbytecode
@Vaibhavdixit02 - I have just sent a pull request in the submission repository:

https://github.com/SciML/GlobalSensitivity.jl/pull/74

Please review the changes carefully and I you are agreed with them, please apply.
