+++ versions.js
@@ -8,5 +8,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v2.1.0";
+var DOCUMENTER_NEWEST = "v2.1.1";
 var DOCUMENTER_STABLE = "stable";
