### 2022-08-15 #4561 @danielskatz
These are related to ISBNs in the bibtex which don't have the minimum of 10 characters.
