### 2022-08-15 #4561 @danielskatz
The error indicates

>!! Invalid Crossref XML v5.3.1 in tmp/b1a2a1cacb0364bf1174279eb83373ce/joss/paper.crossref. Errors: 3.
[76](https://github.com/openjournals/joss-papers/runs/7841093615?check_suite_focus=true#step:6:83)
>  - 116:0: ERROR: Element {http://www.crossref.org/schema/5.3.1}isbn: [facet minLength] The value has a length of 9; this underruns the allowed minimum length of 10.
[77](https://github.com/openjournals/joss-papers/runs/7841093615?check_suite_focus=true#step:6:84)
>  - 147:0: ERROR: Element {http://www.crossref.org/schema/5.3.1}isbn: [facet minLength] The value has a length of 8; this underruns the allowed minimum length of 10.
[78](https://github.com/openjournals/joss-papers/runs/7841093615?check_suite_focus=true#step:6:85)
 > - 208:0: ERROR: Element {http://www.crossref.org/schema/5.3.1}isbn: [facet minLength] The value has a length of 9; this underruns the allowed minimum length of 10.

I suspect these are the line numbers in the bib file: 116, 147, 208 that have the issues, but I'm just guessing
