### 2022-08-15 #4561 @jbytecode
in an entry of bibtex isbn and issn values are same. indeed they are not isbn values

```
isbn = {0951-8320},
issn = {09518320},
```
@Vaibhavdixit02 - could you please correct the isbn values of books?
