### 2022-08-16 #4561 @jbytecode
@Vaibhavdixit02 - Some bibtex entries includes unnecessary isbn items (exactly same with the issn values) are removed. I performed a DOI check for all of them. Please apply the pull request at https://github.com/SciML/GlobalSensitivity.jl/pull/76
