### 2022-08-17 #4561 @kyleniemeyer
Hi @Vaibhavdixit02, it looks like your article is missing a few DOIs, can you add these?
 - DiffernentialEquations.jl paper: 10.5334/jors.151
 - Saltelli et al. (2020: 10.1038/d41586-020-01812-9
