### 2022-08-17 efb6126
add missing DOIs
