+++ joss/paper.bib
@@ -386,6 +386,7 @@ eprint = {
   title={Five ways to ensure that models serve society: {A} manifesto},
   author={Saltelli, Andrea and Bammer, Gabriele and Bruno, Isabelle and Charters, Erica and Di Fiore, Monica and Didier, Emmanuel and Nelson Espeland, Wendy and Kay, John and Lo Piano, Samuele and Mayo, Deborah and others},
   year={2020},
+  doi={10.1038/d41586-020-01812-9},
   publisher={Nature Publishing Group}
 }
 
@@ -434,6 +435,7 @@ abstract = {Models are increasingly being relied upon to inform and support natu
   author={Christopher Rackauckas and Qing Nie},
   journal={Journal of open research software},
   year={2017},
+  doi={10.5334/jors.151},
   volume={5}
 }
 
